# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "Hotel API Documentation",
        "contact": {
            "email": "diogo.coutinho.ads@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get list of hotels",
                "description": "Returns list of hotels",
                "operationId": "37bed2de768fd67fb71c69e2d094276c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HotelResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Create new hotel",
                "description": "Create new hotel and return hotel data",
                "operationId": "11ba64b40696f033635ce936521b42f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HotelResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get hotel by id",
                "description": "Return hotel data",
                "operationId": "e2d52afab1967ac68dd9b600c829788d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hotel to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HotelResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hotel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Update hotel by id",
                "description": "Update hotel data and return updated hotel data",
                "operationId": "d87b62fadb45bf44d27d7c1bd537b658",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hotel to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateHotelFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HotelResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hotel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Delete hotel by id",
                "description": "Delete hotel and return no content",
                "operationId": "ecc4a0c9d0255c6e5428d4ae8e20a680",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hotel to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No content"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hotel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreHotelFormRequest": {
                "title": "StoreHotelFormRequest",
                "required": [
                    "name",
                    "address",
                    "city",
                    "state",
                    "zip_code"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zip_code": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateHotelFormRequest": {
                "required": [
                    "name",
                    "address",
                    "city",
                    "state",
                    "zip_code"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the hotel",
                        "type": "string",
                        "example": "Hotel California"
                    },
                    "address": {
                        "description": "The address of the hotel",
                        "type": "string",
                        "example": "42 Sunset Boulevard"
                    },
                    "city": {
                        "description": "The city where the hotel is located",
                        "type": "string",
                        "example": "Los Angeles"
                    },
                    "state": {
                        "description": "The state where the hotel is located",
                        "type": "string",
                        "example": "California"
                    },
                    "zip_code": {
                        "description": "The zip code of the hotel",
                        "type": "string",
                        "example": "90028"
                    },
                    "website": {
                        "description": "The website of the hotel",
                        "type": "string",
                        "example": "http://www.hotelcalifornia.com",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "HotelResource": {
                "title": "HotelResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zip_code": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "rooms": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoomResource"
                        }
                    }
                },
                "type": "object"
            },
            "RoomResource": {
                "title": "RoomResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}